modules/(SceCtrl/SceTouch): fix and refactor - add controllers connected in log when app boot. - add correct struct of both SceCtrlData(2) - set touch mode in good place. - refactor refresh controllers. - implement sceCtrlGetWirelessControllerInfo. - using DS4 type when is DS4/Dual Sense.
