+++ vita3k/ctrl/include/ctrl/state.h
@@ -17,6 +17,9 @@
 
 #pragma once
 
+#include <ctrl/ctrl.h>
+
+#include <SDL_gamecontroller.h>
 #include <SDL_haptic.h>
 #include <SDL_joystick.h>
 
@@ -28,20 +31,14 @@ struct _SDL_GameController;
 typedef std::shared_ptr<_SDL_GameController> GameControllerPtr;
 typedef std::shared_ptr<_SDL_Haptic> HapticPtr;
 
-enum SceCtrlPadInputMode {
-    SCE_CTRL_MODE_DIGITAL = 0,
-    SCE_CTRL_MODE_ANALOG = 1,
-    SCE_CTRL_MODE_ANALOG_WIDE = 2
-};
-
-enum SceTouchSamplingState {
-    SCE_TOUCH_SAMPLING_STATE_STOP,
-    SCE_TOUCH_SAMPLING_STATE_START
-};
-
 struct Controller {
     GameControllerPtr controller;
     int port;
+};
+
+struct ControllerBinding {
+    SDL_GameControllerButton controller;
+    uint32_t button;
 };
 
 typedef std::map<SDL_JoystickGUID, Controller> ControllerList;
@@ -49,8 +46,8 @@ typedef std::map<SDL_JoystickGUID, Controller> ControllerList;
 struct CtrlState {
     ControllerList controllers;
     int controllers_num = 0;
-    bool free_ports[4] = { true, true, true, true };
+    const char *controllers_name[SCE_CTRL_MAX_WIRELESS_NUM];
+    bool free_ports[SCE_CTRL_MAX_WIRELESS_NUM] = { true, true, true, true };
     SceCtrlPadInputMode input_mode = SCE_CTRL_MODE_DIGITAL;
     SceCtrlPadInputMode input_mode_ext = SCE_CTRL_MODE_DIGITAL;
-    SceTouchSamplingState touch_mode[2] = { SCE_TOUCH_SAMPLING_STATE_STOP, SCE_TOUCH_SAMPLING_STATE_STOP };
 };
