+++ vita3k/touch/CMakeLists.txt
@@ -1,6 +1,8 @@
 add_library(
 	touch
 	STATIC
+	include/touch/functions.h
+	include/touch/state.h
 	include/touch/touch.h
 	src/touch.cpp
 )
